rename files to better indicate they're store modules
