+++ web/package.json
@@ -12,6 +12,7 @@
     "bcryptjs": "^2.4.3",
     "bootstrap": "^4.3.1",
     "jquery": "^3.3.1",
+    "lodash": "^4.17.11",
     "popper.js": "^1.14.7",
     "sjcl": "^1.0.8",
     "vue": "^2.6.6",
