+++ web/src/store/modules/scoreboardModule.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-param-reassign */
 import {
   addCutLineIndicator,
   leaderboardWithTiers,
